feat(dynamic-relays): refactor rcm package
