+++ config/rcm/registry_creator.go
@@ -1,22 +1,12 @@
 package rcm
 
 import (
-	"errors"
 	"fmt"
 
 	"github.com/flashbots/mev-boost/config/relay"
 )
 
-var (
-	ErrConfigProviderFailure        = errors.New("config provider failure")
-	ErrInvalidProposerConfig        = errors.New("invalid proposer config")
-	ErrEmptyBuilderRelays           = errors.New("builder is enabled but has no relays")
-	ErrCannotPopulateProposerRelays = errors.New("cannot populate proposer relays")
-	ErrCannotPopulateDefaultRelays  = errors.New("cannot populate default relays")
-)
-
-// ConfigProvider provider relay configuration.
-type ConfigProvider func() (*relay.Config, error)
+// proposerWalkerFn a helper used for traversing proposal config.
 type proposerWalkerFn func(publicKey relay.ValidatorPublicKey, cfg relay.Relay) error
 
 // RegistryCreator creates Relay Registries.
